Provide more flexible structure for work package contract validations
