+++ spec/lib/api/v3/work_packages/work_package_contract_spec.rb
@@ -28,7 +28,7 @@
 
 require 'spec_helper'
 
-describe ::API::V3::WorkPackages::WorkPackageContract do
+describe ::API::V3::WorkPackages::BaseContract do
   let(:work_package) do
     FactoryGirl.create(:work_package,
                        done_ratio: 50,
